+++ src/components/Plugins/DynamicText.tsx
@@ -44,7 +44,7 @@ export const DynamicText: React.FC = () => {
         </LinkButton>
         <LinkButton
           variant="secondary"
-          icon="star"
+          icon="github"
           target="_blank"
           href="https://github.com/volkovlabs/volkovlabs-dynamictext-panel"
         >