+++ src/components/Plugins/GrafanaAPI.tsx
@@ -38,7 +38,7 @@ export const GrafanaAPI: React.FC = () => {
         </LinkButton>
         <LinkButton
           variant="secondary"
-          icon="star"
+          icon="github"
           target="_blank"
           href="https://github.com/volkovlabs/volkovlabs-grapi-datasource"
         >