+++ src/components/Plugins/Static.tsx
@@ -39,7 +39,7 @@ export const Static: React.FC = () => {
         </LinkButton>
         <LinkButton
           variant="secondary"
-          icon="star"
+          icon="github"
           target="_blank"
           href="https://github.com/volkovlabs/volkovlabs-static-datasource"
         >